+++ docs/victoriametrics/changelog/CHANGELOG.md
@@ -36,6 +36,8 @@ See also [LTS releases](https://docs.victoriametrics.com/victoriametrics/lts-rel
 
 * BUGFIX: [vmui](https://docs.victoriametrics.com/victoriametrics/single-server-victoriametrics/#vmui): fix escaping for label names with special characters. See [#10485](https://github.com/VictoriaMetrics/VictoriaMetrics/issues/10485).
 
+* FEATURE: [OpenTelemetry](https://docs.victoriametrics.com/victoriametrics/single-server-victoriametrics/#sending-data-via-opentelemetry): add a `-opentelemetry.usePermissiveLabelSanitization` flag to allow labels starting with a single underscore to pass through without being prefixed when `-opentelemetry.usePrometheusNaming` is active. See [#9663](https://github.com/VictoriaMetrics/VictoriaMetrics/issues/9663).
+
 ## [v1.136.0](https://github.com/VictoriaMetrics/VictoriaMetrics/releases/tag/v1.136.0)
 
 Released at 2026-02-13